+++ CHANGELOG.md
@@ -3,10 +3,14 @@
 
 ## Table of Contents
 
+* [Release 1.6.1](#release-161)
 * [Release 1.5.1](#release-151)
 ---
 
+### Release 1.6.1
+
+1. Renamed class TS_ILI9341_map to simply TS_ILI9341, with .h/.cpp files also renamed, and added enum eTouchEvent and class functions getTouchEvent() and setTouchReleaseParams(), to support touchscreen touch and release events.
+
 ### Release 1.5.1
 
-1. Initial release of this library, cloned from version 1.4 of Paul Stoffregen's
-XPT2046_Touchscreen library at https://github.com/PaulStoffregen/XPT2046_Touchscreen.
+1. Initial release of this library, cloned from version 1.4 of Paul Stoffregen's XPT2046_Touchscreen library at https://github.com/PaulStoffregen/XPT2046_Touchscreen.